+++ src/table/index.js
@@ -1,5 +1,8 @@
 // @flow
+import { type PagedTableProps } from "./paged-table";
+
 export { PagedTable } from "./paged-table";
 export { SimpleTable } from "./simple-table";
 export { LargeSimpleTable } from "./large-simple-table";
 export { ExportTableButton } from "./components/export-button";
+export type { PagedTableProps };